==> notebooks/3.recap_all.ipynb <==
import os

print(os.path.abspath(os.curdir))

import os
os.chdir("..")

print(os.path.abspath(os.curdir))

import os
import pandas as pd
baseline_path = os.path.join('experiments', 'recap', 'baseline_15', 'recap_overall_baseline_15.xlsx')
multitask_path = os.path.join('experiments', 'recap', 'multitask_15', 'recap_overall_multitask_15.xlsx')
recap_baseline = pd.read_excel(baseline_path)
recap_multitask = pd.read_excel(multitask_path)
recap_all = pd.concat([recap_baseline, recap_multitask], ignore_index=True)

output_dir = os.path.join('experiments', 'recap')
recap_all.to_excel(os.path.join(output_dir, 'final-recap-all.xlsx'), index=False)

#### Recap the best models

best_baseline_path = os.path.join('experiments', 'recap', 'baseline_15', 'best_model_acc_f1_baseline_15.xlsx')
best_multitask_path = os.path.join('experiments', 'recap', 'multitask_15', 'best_model_acc_f1_multitask_15.xlsx')
recap_best_baseline = pd.read_excel(best_baseline_path)
recap_best_multitask = pd.read_excel(best_multitask_path)
recap_best_all = pd.concat([recap_best_baseline, recap_best_multitask], ignore_index=True)

output_dir = os.path.join('experiments', 'recap')
recap_best_all.to_excel(os.path.join(output_dir, 'recap-best-all.xlsx'), index=False)
