==> notebooks/4.best_model.ipynb <==
import os

print(os.path.abspath(os.curdir))

import os
os.chdir("..")

print(os.path.abspath(os.curdir))

#### Filtered Dataset

import os
import pandas as pd


def build_recap_best_model(scenario, workdir):
    recap_path = os.path.join(workdir, f'recap_overall_{scenario}.xlsx')
    recap_df = pd.read_excel(recap_path)
    recap_filtered = recap_df[recap_df['dataset'] == 'filtered']
    group_columns = ['encoder', 'scenario']
    value_columns = ['accuracy', 'f1_score']

    # Use groupby to group the DataFrame by the specified columns
    # Then apply idxmax() to get the index of the maximum value in each group
    max_indices = recap_filtered.groupby(group_columns)['accuracy'].idxmax()
    best_model_acc_f1_filtered = recap_filtered.sort_values(value_columns, ascending=False).groupby(group_columns).head(1)
    # Use loc to retrieve the corresponding rows based on the indices
    max_rows_filtered = recap_filtered.loc[max_indices]
    
    recap_unfiltered = recap_df[recap_df['dataset'] == 'unfiltered']
    # group_columns = ['encoder', 'scenario']
    # value_column = 'accuracy'

    # Use groupby to group the DataFrame by the specified columns
    # Then apply idxmax() to get the index of the maximum value in each group
    max_indices = recap_unfiltered.groupby(group_columns)['accuracy'].idxmax()
    best_model_acc_f1_unfiltered = recap_unfiltered.sort_values(value_columns, ascending=False).groupby(group_columns).head(1)

    # Use loc to retrieve the corresponding rows based on the indices
    max_rows_unfiltered = recap_unfiltered.loc[max_indices]
    
    best_model = pd.concat([max_rows_filtered, max_rows_unfiltered], ignore_index=True)
    best_model.to_excel(os.path.join(workdir, f'best_model_acc_{scenario}.xlsx'), index=False)
    
    best_model_acc_f1 = pd.concat([best_model_acc_f1_filtered, best_model_acc_f1_unfiltered], ignore_index=True)
    best_model_acc_f1.to_excel(os.path.join(workdir, f'best_model_acc_f1_{scenario}.xlsx'), index=False)

scenario = 'baseline_15'
# source_dir = os.path.join('experiments', scenario)
workdir = os.path.join('experiments', 'recap', scenario)
build_recap_best_model(scenario, workdir)

scenario = 'multitask_15'
# source_dir = os.path.join('experiments', scenario)
workdir = os.path.join('experiments', 'recap', scenario)
build_recap_best_model(scenario, workdir)
