==> notebooks/6.hyperparam.ipynb <==
import os

print(os.path.abspath(os.curdir))

import os
os.chdir("..")

print(os.path.abspath(os.curdir))

# Ablation Study
This ablation study is based on the chi-square test results. Only significant hyperparameters toward the accuracy are analyzed.

import pandas as pd

baseline_path = 'experiments/recap/final-recap-all.xlsx'
baseline_df = pd.read_excel(baseline_path)
baseline_df.head(5)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
grouped_acc = baseline_df.groupby(['dataset', 'encoder', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_acc

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
grouped_f1 = baseline_df.groupby(['dataset', 'encoder', 'scenario'])['f1_score'].agg(['mean', 'std'])
grouped_f1

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
grouped_conf = baseline_df.groupby(['dataset', 'encoder', 'scenario'])['confidence_avg'].agg(['mean', 'std'])
grouped_conf

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
grouped_epoch = baseline_df.groupby(['dataset', 'encoder', 'scenario'])['best_epoch'].agg(['mean', 'std'])
grouped_epoch

# Pivot the grouped data to create the desired format
pivot_table = grouped.reset_index().pivot(index=['dataset', 'scenario'], columns='encoder', values=['mean', 'std'])
print(pivot_table.columns)
# Flatten the column names in the resulting DataFrame
pivot_table.columns = ['{}_{}'.format(col[0], col[1]) for col in pivot_table.columns]

# Reset the index to make 'dataset' and 'scenario' regular columns
final_df = pivot_table.reset_index()

# Rename columns as needed
final_df.rename(columns={'dataset': 'Dataset', 'scenario': 'Scenario'}, inplace=True)

# Display the final DataFrame
# final_df.to_excel
root_workdir = os.path.join('experiments', 'recap', 'hyperparam')
if not os.path.exists(root_workdir):
    os.makedirs(root_workdir)
final_df.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_per_scenario.xlsx'), index=False, float_format='%.3f')
print(final_df)

## 1. Filtered Dataset

filtered_df = baseline_df[baseline_df['dataset'] == 'filtered']
filtered_df

### A. F-GRU (scenario-class weight)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
filtered_gru = filtered_df[filtered_df['encoder'] == 'gru']
grouped_f_gru = filtered_gru.groupby(['class_weight', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_f_gru.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_f_gru_scenario_class-weight.xlsx'), index=True, float_format='%.3f')
grouped_f_gru

### B. F-LSTM (scenario-bidirectionality)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
filtered_lstm = filtered_df[filtered_df['encoder'] == 'lstm']
grouped_f_lstm = filtered_lstm.groupby(['bidirectional', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_f_lstm.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_f_lstm_scenario_bidirectionality.xlsx'), index=True, float_format='%.3f')
grouped_f_lstm

### C. F-Linear (scenario-class weight)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
filtered_linear = filtered_df[filtered_df['encoder'] == 'linear']
grouped_f_linear = filtered_linear.groupby(['class_weight', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_f_linear.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_f_linear_scenario_class-weight.xlsx'), index=True, float_format='%.3f')
grouped_f_linear

### D. F-Transformer (scenario-class weight)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
filtered_transformer = filtered_df[filtered_df['encoder'] == 'transformer']
grouped_f_trans_class = filtered_transformer.groupby(['class_weight', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_f_trans_class.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_f_trans_scenario_class-weight.xlsx'), index=True, float_format='%.3f')
grouped_f_trans_class

### E. F-Transformer (scenario - n_layer)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
filtered_transformer = filtered_df[filtered_df['encoder'] == 'transformer']
grouped_f_trans_nlayer = filtered_transformer.groupby(['n_layer', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_f_trans_nlayer.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_f_trans_scenario_n-layer.xlsx'), index=True, float_format='%.3f')
grouped_f_trans_nlayer

### F. F-Transformer (scenario-n_layer)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
filtered_transformer = filtered_df[filtered_df['encoder'] == 'transformer']
grouped_f_trans_layer = filtered_transformer.groupby(['n_layer', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_f_trans_layer.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_f_trans_scenario_n-layer.xlsx'), index=True, float_format='%.3f')
grouped_f_trans_layer

## 2. Unfiltered Dataset

unfiltered_df = baseline_df[baseline_df['dataset'] == 'unfiltered']
unfiltered_df

### A. U-GRU (scenario - n_layer)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
unfiltered_gru = unfiltered_df[unfiltered_df['encoder'] == 'gru']
grouped_u_gru_nlayer = unfiltered_gru.groupby(['n_layer', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_u_gru_nlayer.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_u_gru_scenario_n-layer.xlsx'), index=True, float_format='%.3f')
grouped_u_gru_nlayer

### A-2. U-GRU (scenario - class weight)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
unfiltered_gru = unfiltered_df[unfiltered_df['encoder'] == 'gru']
grouped_u_gru_class = unfiltered_gru.groupby(['class_weight', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_u_gru_class.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_u_gru_scenario_class-weight.xlsx'), index=True, float_format='%.3f')
grouped_u_gru_class

### B. U-LSTM (scenario - n_layer)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
unfiltered_lstm = unfiltered_df[unfiltered_df['encoder'] == 'lstm']
grouped_u_lstm_nlayer = unfiltered_lstm.groupby(['n_layer', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_u_lstm_nlayer.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_u_lstm_scenario_n-layer.xlsx'), index=True, float_format='%.3f')
grouped_u_lstm_nlayer

### C. U-Linear (scenario - pooling)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
unfiltered_linear = unfiltered_df[unfiltered_df['encoder'] == 'linear']
grouped_u_linear_pooling = unfiltered_linear.groupby(['pooling', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_u_linear_pooling.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_u_linear_scenario_pooling.xlsx'), index=True, float_format='%.3f')
grouped_u_linear_pooling

### D-1. U-Transformer (scenario-class weight)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
unfiltered_transformer = unfiltered_df[unfiltered_df['encoder'] == 'transformer']
grouped_u_trans_class = unfiltered_transformer.groupby(['class_weight', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_u_trans_class.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_u_trans_scenario_class-weight.xlsx'), index=True, float_format='%.3f')
grouped_u_trans_class

### D-2. U-Transformer (scenario-n_layer)

# Group by 'dataset', 'encoder', and 'scenario', and calculate mean and std
# filtered_transformer = unfiltered_df[unfiltered_df['encoder'] == 'transformer']
grouped_u_trans_layer = unfiltered_transformer.groupby(['n_layer', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_u_trans_layer.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_u_trans_scenario_n-layer.xlsx'), index=True, float_format='%.3f')
grouped_u_trans_layer

## 3. Three-paired Analysis

### A. F-Transformer (n_layer - class weight - scenario)

grouped_f_trans_3 = filtered_transformer.groupby(['n_layer', 'class_weight', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_f_trans_3.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_f_trans_scenario_class-weight_n-layer.xlsx'), index=True, float_format='%.3f')
grouped_f_trans_3

### B. U-Transformer (n_layer - class weight - scenario)

grouped_u_trans_3 = unfiltered_transformer.groupby(['n_layer', 'class_weight', 'scenario'])['accuracy'].agg(['mean', 'std'])
grouped_u_trans_3.to_excel(os.path.join('experiments', 'recap', 'hyperparam', 'recap_u_trans_scenario_class-weight_n-layer.xlsx'), index=True, float_format='%.3f')
grouped_u_trans_3
